Moved from youtube-dl to pytube
Although youtube-dl is a great tool, it only makes sense in the
context of a shell command line or a bash script. Called from a
python script it just fills the standard output with garbage. It's
quite bad, since the threading mixed all of the youtube-dl outputs
from the different songs.

The interface of pytube is quite lower level, but even in its
current, not very curated state it provides decent functionality.

TODO: On a short term, introduce "easy" callbacks to the
downloaders, so we know when songs have been downloaded, f.x.
On a longer term, unify those in a single callback so the
information is clear and organized. Could, for example, include
the amount of songs downloaded/remaining.