Included stream as an atr. in Downloader
The stream didn't have much relevance before, so it was accessed
implicitly through the select_stream function, which was called
from another function.

min_duration_video has been taken out of the class and now the
funciton select_stream is called from __init__, so stream can be
saved as a Downloader attribute.

stream needs to be a Downloader attribute so a list of streams can
be computed for initialization of the ProgressBar instances.