scrap_songs now returns artist names as well
The queries will now be more specific, since previously it was
impossible to download a cover unless it was more popular than the
original.

This being said, the contents of scrap.py are still quite messy. The
module still needs to migrate from using regex to parsing the html;
besides not being a very readable approach, currently duplicated
matches have to be deleted manually (since the order of appearance
has to be preserved). There shouldn't be duplications from the
beginning.