Conversor warns if file already exists
Currently only a print statement is added. In the future, this
type of message should be done through a logger.

Also, Converter has some dependencies to os, which could be
substituted by simple Path methods' calls.

A logger should be implemented if more "warning prints" are needed;
removing the dependencies to os should be easy, and the code would
benefit from it in simplicity.