Split the test_suite into individual files
Each module has its corresponding test file (except for the
progressbar module, which doesn't have any tests associated).

Since all the tests are gathered at the testing folder, importing
has become clumsier. Currently the path is added manually to path
in sys, but the solution is ugly (and pylint is complaining). A
better solution needs to be found.